Renamed modules according to Java conventions, some code was rewritten without changing behaviour and implemented Singleton Pattern in MainMenu
