+++ src/memPackage/HighScoreMenu.java
@@ -13,18 +13,18 @@ import javax.swing.table.DefaultTableModel;
  *
  * @author avih2
  */
-public class highscoreMenu extends javax.swing.JFrame {
+public class HighScoreMenu extends javax.swing.JFrame {
 
     /**
      * Creates new form highscoreMenu
      */
-    public highscoreMenu() {
+    public HighScoreMenu() {
         initComponents();
         Color colour = new Color(172, 241, 250);
         getContentPane().setBackground(colour);
         
-        String username = instructionsMenu.getUsername();
-        int score = gameMenu.getScore();
+        String username = InstructionsMenu.getUsername();
+        int score = GameMenu.getScore();
         DefaultTableModel model = (DefaultTableModel) jTable1.getModel();
         if(score!=0){
             Object[] row = {username, score};
@@ -135,8 +135,7 @@ public class highscoreMenu extends javax.swing.JFrame {
     private void returnButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_returnButtonActionPerformed
         // TODO add your handling code here:
         close();
-        mainMenu menu = new mainMenu();
-        menu.setVisible(true);
+        MainMenu.showMenu();
     }//GEN-LAST:event_returnButtonActionPerformed
 
     /**
@@ -156,20 +155,20 @@ public class highscoreMenu extends javax.swing.JFrame {
                 }
             }
         } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(highscoreMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(HighScoreMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(highscoreMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(HighScoreMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(highscoreMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(HighScoreMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(highscoreMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(HighScoreMenu.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
         //</editor-fold>
         
         /* Create and display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
             public void run() {
-                new highscoreMenu().setVisible(true);
+                new HighScoreMenu().setVisible(true);
                 
             }
         });
